+++ Changelog.md
@@ -67,6 +67,7 @@ With the port to Bootstrap 3, app/views/terms/default.haml has a new structure. 
 * Display username and avatar for NSFW posts in mobile view [#6245](https://github.com/diaspora/diaspora/6245)
 * Prevent multiple comment boxes on mobile [#6363](https://github.com/diaspora/diaspora/pull/6363)
 * Correctly display location in post preview [#6429](https://github.com/diaspora/diaspora/pull/6429)
+* Do not fail when submitting an empty comment in the mobile view [#6543](https://github.com/diaspora/diaspora/pull/6543)
 
 ## Features
 * Support color themes [#6033](https://github.com/diaspora/diaspora/pull/6033)